+++ src/main/java/controller/extra/ViewToExerciseController.java
@@ -1,7 +1,5 @@
 package controller.extra;
 
-import java.util.ArrayList;
-
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -9,8 +7,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import controller.Controller;
-import model.ToExercise;
-import model.service.StaticManager;
 
 public class ViewToExerciseController implements Controller {
 	private static final Logger log = LoggerFactory.getLogger(ViewToExerciseController.class);
@@ -19,14 +15,14 @@ public class ViewToExerciseController implements Controller {
 	public String execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
 		int exerciserId = Integer.parseInt(request.getParameter("exerciserId"));
 
-		StaticManager manager = StaticManager.getInstance();
-
-		ArrayList<ToExercise> list = manager.findToExercise(exerciserId);
-
-		if (list != null) {
-			request.setAttribute("ToExerciseList", list);
-			return "redirect:/mypage/ToExercise";
-		}
+//		StaticManager manager = StaticManager.getInstance();
+//
+//		ArrayList<ToExercise> list = manager.findToExercise(exerciserId);
+//
+//		if (list != null) {
+//			request.setAttribute("ToExerciseList", list);
+//			return "redirect:/mypage/ToExercise";
+//		}
 
 		return " /mypage/ToExercise.jsp";
 	}
