+++ src/main/java/controller/myPage/StaticController.java
@@ -1,0 +1,5 @@
+package controller.myPage;
+
+public class StaticController {
+
+}
