+++ src/main/java/model/dao/StaticDao.java
@@ -2,11 +2,9 @@ package model.dao;
 
 import java.sql.ResultSet;
 import java.util.ArrayList;
-import java.util.Date;
 
 import model.CompareInbody;
 import model.CompareStatic;
-import model.ToExercise;
 
 public class StaticDao {
 	private JDBCUtil jdbcUtil = null; // JDBCUtil 참조 변수 선언
@@ -14,104 +12,6 @@ public class StaticDao {
 	public StaticDao() { // 생성자
 		jdbcUtil = new JDBCUtil(); // JDBCUtil 객체 생성 및 활용
 
-	}
-
-	public ArrayList<ToExercise> findToExercise(int exerciserId) {
-		ArrayList<ToExercise> toExerciseList = new ArrayList<ToExercise>();
-		ToExercise list;
-
-		String query = "select * from toexercise where exerciserId=?";
-		Object[] param = new Object[] { exerciserId };
-		jdbcUtil.setSqlAndParameters(query, param);
-
-		try {
-			ResultSet rs = jdbcUtil.executeQuery();
-
-			while (rs.next()) {
-				int itemid = rs.getInt("itemid");
-				String creationDate = rs.getString("creationdate");
-				String content = rs.getString("content");
-				String checked = rs.getString("checked");
-
-				list = new ToExercise(itemid, creationDate, content, checked, exerciserId);
-				toExerciseList.add(list);
-			}
-			return toExerciseList;
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			jdbcUtil.close();
-		}
-		return null;
-	}
-
-	public ArrayList<ToExercise> addToExercise(int exerciserId, String content) {
-		Date date = new Date();
-		java.util.Date sqlDate = new java.sql.Date(date.getTime());
-
-		String query = "insert into toexercise  (creationdate, content, checked, exerciserid)\r\n"
-				+ "values (?, ?, F , ?)";
-		Object[] param = new Object[] { date, content, exerciserId };
-		jdbcUtil.setSqlAndParameters(query, param);
-
-		try {
-			int result = jdbcUtil.executeUpdate(); // insert 문 실행
-			if (result == 1) {
-				return findToExercise(exerciserId); // insert 에 의해 반영된 레코드 수 반환
-			}
-		} catch (Exception e) {
-			jdbcUtil.rollback();
-			e.printStackTrace();
-		} finally {
-			jdbcUtil.commit();
-			jdbcUtil.close(); // ResultSet, PreparedStatement, Connection 반환
-		}
-
-		return null;
-	}
-
-	public ArrayList<ToExercise> checkToExercise(int exerciserId, int itemId) {
-		String query = "UPDATE toexercise " + "SET checked = ? " + " WHERE exerciserid = ? and itemdId = ? ";
-
-		Object[] param = new Object[] { "T", exerciserId, itemId };
-
-		jdbcUtil.setSqlAndParameters(query, param);
-
-		try {
-			int result = jdbcUtil.executeUpdate(); // insert 문 실행
-			if (result == 1) {
-				return findToExercise(exerciserId); // insert 에 의해 반영된 레코드 수 반환
-			}
-		} catch (Exception e) {
-			jdbcUtil.rollback();
-			e.printStackTrace();
-		} finally {
-			jdbcUtil.commit();
-			jdbcUtil.close(); // ResultSet, PreparedStatement, Connection 반환
-		}
-
-		return null;
-	}
-
-	public ArrayList<ToExercise> deleteToExercise(int exerciserId, int itemId) {
-		String query = "DELETE FROM toexercise WHERE exerciserid = ? and itemdId = ? ";
-		Object[] param = new Object[] { exerciserId, itemId };
-		jdbcUtil.setSqlAndParameters(query, param);
-
-		try {
-			int result = jdbcUtil.executeUpdate(); // insert 문 실행
-			if (result == 1) {
-				return findToExercise(exerciserId); // insert 에 의해 반영된 레코드 수 반환
-			}
-		} catch (Exception e) {
-			jdbcUtil.rollback();
-			e.printStackTrace();
-		} finally {
-			jdbcUtil.commit();
-			jdbcUtil.close(); // ResultSet, PreparedStatement, Connection 반환
-		}
-
-		return null;
 	}
 
 	public CompareStatic calculateStatic(int exerciserId) {
