+++ src/main/java/model/service/StaticManager.java
@@ -1,9 +1,6 @@
 package model.service;
 
-import java.util.ArrayList;
-
 import model.CompareStatic;
-import model.ToExercise;
 import model.dao.StaticDao;
 
 public class StaticManager {
@@ -20,22 +17,6 @@ public class StaticManager {
 
 	public static StaticManager getInstance() {
 		return manager;
-	}
-
-	public ArrayList<ToExercise> findToExercise(int exerciserId) {
-		return staticDao.findToExercise(exerciserId);
-	}
-
-	public ArrayList<ToExercise> addToExercise(int exerciserId, String content) {
-		return staticDao.addToExercise(exerciserId, content);
-	}
-
-	public ArrayList<ToExercise> checkToExercise(int exerciserId, int itemId) {
-		return staticDao.checkToExercise(exerciserId, itemId);
-	}
-
-	public ArrayList<ToExercise> deleteToExercise(int exerciserId, int itemId) {
-		return staticDao.deleteToExercise(exerciserId, itemId);
 	}
 
 	public CompareStatic calculateStatic(int exerciserId) {
