+++ src/main/webapp/WEB-INF/myPage/static.jsp
@@ -1,0 +1,16 @@
+<%@ page language="java" contentType="text/html; charset=utf-8"
+    pageEncoding="utf-8"%>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
+<!DOCTYPE html>
+<html>
+<head>
+	<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+	<title>static</title>
+    <link rel="stylesheet" href="<c:url value='/css/frame.css' />" type="text/css">
+    <link rel="stylesheet" href="<c:url value='/css/static.css' />" type="text/css">
+</head>
+
+<%@ include file="../frameHeader.jsp" %>
+
+
+<%@ include file="../frameFooter.jsp" %>
