+++ src/main/webapp/css/static.css
@@ -1,0 +1,101 @@
+@charset "utf-8";
+/* main */
+#main {
+    box-sizing: border-box;
+    padding: 30px 30px 0 30px;
+
+    height: auto;
+    background: #FFFFFF;
+    border: 1px solid #909090;
+}
+
+/*서브 제목*/
+#subTitle .subTitle{
+    position: relative;
+    font-style: normal;
+    font-weight: 600;
+    font-size: 22px;
+
+    margin-bottom: 20px;
+    color: #414141;
+}
+
+.flex-container2 {
+    display: flex;
+    flex-direction: column;
+}
+
+.flex-nickname {
+    display: flex;
+    justify-content: flex-start;
+    align-items: baseline;
+    padding-bottom: 15px;
+    padding-left: 8px;
+
+    font-weight: 600;
+    font-size: 23px;
+    color: #414141;
+}
+
+#님 {
+    font-weight: 400;
+    font-size: 15px;
+    color: #414141;
+}
+
+#icon_bottom_line {
+    border-bottom: 1px solid #C6C6C6;
+    padding-bottom: 15px;
+}
+
+.flex-container-icon{
+    display: flex;
+    flex-direction: row;
+    padding: 10px;
+    margin-bottom: 10px;
+    justify-content: space-between;
+    align-items: center;
+    border-bottom: 1px solid #C6C6C6;
+}
+
+.flex-container-icon2{
+    display: flex;
+    flex-direction: row;
+    padding: 10px;
+    margin-bottom: 10px;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.flex-container-icon-sub1{
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+}
+
+.category2 {
+    font-weight: 600;
+    font-size: 16px;
+    color: #414141;
+}
+
+.state{
+    font-weight: 600;
+    font-size: 18px;
+    color: #414141;
+}
+
+button {
+    background-color: #FFFFFF;
+    border: none;
+    font-family: 'Noto Sans KR';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 12px;
+    line-height: 23px;
+    text-align: center;
+    text-decoration-line: underline;
+    color: #909090;
+
+}
